+++ resources/views/layout/customer/type/input.blade.php
@@ -22,7 +22,7 @@
                 <div class="card-body">
                     <div class="mb-3">
                         <label class="form-label">Name<span class="text-danger">*</span></label>
-                        <input type="text" name="name" class="form-control w-50 @error('name') is-invalid @enderror" value="{{ old('name', $data->name) }}" placeholder="Nama Tipe Konsumen">
+                        <input type="text" name="name" class="form-control w-50 @error('name') is-invalid @enderror" value="{{ old('name', $data->name) }}" placeholder="Nama Tipe Konsumen" validate>
                         @error('name')
                             <div class="invalid-feedback">
                                 {{ $message }}
@@ -42,4 +42,10 @@
         </div>
     </div>
     @include('form.header.end')
+@endsection
+
+@section('script')
+    <script>
+        initInput('{{ $config['url'] }}');
+    </script>
 @endsection